Extracted SfFilter from Vue file to specific files
