+++ packages/vue/src/components/molecules/SfFilter/SfFilter.html
@@ -1,0 +1,3 @@
+<div class="sf-filter">
+  <slot />
+</div>
